feat: reorganizing sections in a folder and switching sections-container to app.css
