+++ frontend/src/assets/css/App.css
@@ -48,3 +48,14 @@
   letter-spacing: .5px;
   border-color: var(--main-blue);
 }
+
+/*botão que separa as seções*/
+.section-button {
+  font-size: 14px;
+  background-color: rgba(13,13,13,100%);
+  color: rgba(255,255,255, 70%);
+  border: 1px solid rgba(128, 128, 128, 0.5);
+  margin: 10px 0 25px ;
+  cursor: text;
+  color: rgba(255, 255, 255, 70%);
+}
