+++ frontend/src/pages/Home.jsx
@@ -1,14 +1,13 @@
-import Duvidas from "../components/Duvidas";
-import Plans from "../components/Plans";
+import Duvidas from "../sections/Duvidas";
+import Plans from "../sections/Plans";
+import Contrate from "../sections/Contrate"
 import './Home.css'
 
 
 export default function Home() {
   return (
     <div className="home">
-
       <Plans />
-
       <Duvidas />
       <Contrate/>
     </div>
