+++ frontend/src/sections/Contrate.css
@@ -1,3 +1,11 @@
+.outer{
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
 .contrate-container{
   width: 75%;
   display: flex;
