+++ frontend/src/sections/Duvidas.css
@@ -6,16 +6,6 @@
     transition: 0.6s ease-out, transform 0.6s ease-out;
 }
 
-.duvidas-button {
-    font-size: 14px;
-    background-color: rgba(13,13,13,100%);
-    color: rgba(255,255,255, 70%);
-    border: 1px solid rgba(128, 128, 128, 0.5);
-    margin-bottom: 25px;
-    cursor: text;
-    color: rgba(255, 255, 255, 70%);
-}
-
 
 .duvidas-container,
 .duvidas-container * {
